Move TelephoneAdvisoryReportsNumberCard to components/index/cards
